movie_db_picker/requirements.txt
django
django-widget-tweaks
djangorestframework
requests